+++ app/actions.go
@@ -26,19 +26,19 @@ import (
 
 	log "github.com/golang/glog"
 	"github.com/megamsys/gulp/handlers"
-	//"github.com/megamsys/gulp/state/provisioner/chefsolo"
+	"github.com/megamsys/gulp/state/provisioner/chefsolo"
 	"github.com/megamsys/libgo/action"
 	"github.com/megamsys/libgo/exec"
 	"github.com/tsuru/config"
 )
 
-func TorpedoCommandExecutor(command string, app *handlers.DeepAssembly) (action.Result, error) {
+func TorpedoCommandExecutor(command string, app *handlers.AssemblyWithComponents) (action.Result, error) {
 	var e exec.OsExecutor
 	var commandWords []string
 
 	ctype := strings.Split(app.ToscaType, ".")
 	commandWords = strings.Fields(command + " " + ctype[2])
-	log.Debug("Command Executor entry: %s\n", app)
+	log.Info("Command Executor entry: %s\n", app)
 	megam_home, ckberr := config.GetString("megam_home")
 	if ckberr != nil {
 		return nil, ckberr
@@ -70,8 +70,8 @@ func TorpedoCommandExecutor(command string, app *handlers.DeepAssembly) (action.
 
 	foutwriter := bufio.NewWriter(fout)
 	ferrwriter := bufio.NewWriter(ferr)
-	log.Debug(commandWords)
-	log.Debug("Length: %s", len(commandWords))
+	log.Info(commandWords)
+	log.Info("Length: %s", len(commandWords))
 
 	defer ferrwriter.Flush()
 	defer foutwriter.Flush()
@@ -139,7 +139,7 @@ func ChefCommandExecutor(commandWords []string, app *chefsolo.Provisioner) (acti
 	return &app, nil
 }
 
-func CommandExecutor(command string, app *handlers.DeepAssembly) (action.Result, error) {
+func CommandExecutor(command string, app *handlers.AssemblyWithComponents) (action.Result, error) {
 
 	for i := range app.Components {
 		ctype := strings.Split(app.Components[i].ToscaType, ".")
@@ -156,6 +156,113 @@ func CommandExecutor(command string, app *handlers.DeepAssembly) (action.Result,
 
 	return &app, nil
 }
+
+func ComponentCommandExecutor(app *handlers.Component) (action.Result, error) {
+	var e exec.OsExecutor
+	var commandWords []string
+
+	commandWords = strings.Fields(app.Command)
+	log.Info("Command Executor entry: %s\n", app)
+	megam_home, ckberr := config.GetString("megam_home")
+	if ckberr != nil {
+		return nil, ckberr
+	}
+	appName := app.Name
+	basePath := megam_home + "logs"
+	dir := path.Join(basePath, appName)
+
+	fileOutPath := path.Join(dir, appName+"_out")
+	fileErrPath := path.Join(dir, appName+"_err")
+	if _, err := os.Stat(dir); os.IsNotExist(err) {
+		log.Info("Creating directory: %s\n", dir)
+		if errm := os.MkdirAll(dir, 0777); errm != nil {
+			return nil, errm
+		}
+	}
+	// open output file
+	fout, outerr := os.OpenFile(fileOutPath, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+	if outerr != nil {
+		return nil, outerr
+	}
+	defer fout.Close()
+	// open Error file
+	ferr, errerr := os.OpenFile(fileErrPath, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+	if errerr != nil {
+		return nil, errerr
+	}
+	defer ferr.Close()
+
+	foutwriter := bufio.NewWriter(fout)
+	ferrwriter := bufio.NewWriter(ferr)
+	log.Info(commandWords)
+	log.Info("Length: %s", len(commandWords))
+
+	defer ferrwriter.Flush()
+	defer foutwriter.Flush()
+
+	if len(commandWords) > 0 {
+		if err := e.Execute(commandWords[0], commandWords[1:len(commandWords)], nil, foutwriter, ferrwriter); err != nil {
+			fmt.Println(err)
+			return nil, err
+		}
+	}
+
+	return &app, nil
+}
+
+
+func RedeployCommandExecutor(app *handlers.Component) (action.Result, error) {
+	var e exec.OsExecutor
+	var commandWords []string
+
+	commandWords = strings.Fields(app.Command)
+	log.Info("Command Executor entry: %s\n", app)
+	megam_home, ckberr := config.GetString("MEGAM_HOME")
+	if ckberr != nil {
+		return nil, ckberr
+	}
+	appName := app.Name
+	basePath := megam_home + "logs"
+	dir := path.Join(basePath, appName)
+
+	fileOutPath := path.Join(dir, appName+"_out")
+	fileErrPath := path.Join(dir, appName+"_err")
+	if _, err := os.Stat(dir); os.IsNotExist(err) {
+		log.Info("Creating directory: %s\n", dir)
+		if errm := os.MkdirAll(dir, 0777); errm != nil {
+			return nil, errm
+		}
+	}
+	// open output file
+	fout, outerr := os.OpenFile(fileOutPath, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+	if outerr != nil {
+		return nil, outerr
+	}
+	defer fout.Close()
+	// open Error file
+	ferr, errerr := os.OpenFile(fileErrPath, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+	if errerr != nil {
+		return nil, errerr
+	}
+	defer ferr.Close()
+
+	foutwriter := bufio.NewWriter(fout)
+	ferrwriter := bufio.NewWriter(ferr)
+	log.Info(commandWords)
+	log.Info("Length: %s", len(commandWords))
+
+	defer ferrwriter.Flush()
+	defer foutwriter.Flush()
+
+	if len(commandWords) > 0 {
+		if err := e.Execute(commandWords[0], commandWords[1:len(commandWords)], nil, foutwriter, ferrwriter); err != nil {
+			fmt.Println(err)
+			return nil, err
+		}
+	}
+	return &app, nil
+}
+
 
 /**
 ** state up the virtual machine
@@ -180,17 +287,39 @@ var stateup = action.Action{
 	MinParams: 1,
 }
 
+
+var startApp = action.Action{
+	Name: "startapp",
+	Forward: func(ctx action.FWContext) (action.Result, error) {
+		var app handlers.AssemblyWithComponents
+		switch ctx.Params[0].(type) {
+		case handlers.AssemblyWithComponents:
+			app = ctx.Params[0].(handlers.AssemblyWithComponents)
+		case *handlers.AssemblyWithComponents:
+			app = *ctx.Params[0].(*handlers.AssemblyWithComponents)
+		default:
+			return nil, errors.New("First parameter must be App or *handlers.AssemblyWithComponents.")
+		}
+		return CommandExecutor("start", &app)
+	},
+	Backward: func(ctx action.BWContext) {
+		log.Info("[%s] Nothing to recover")
+	},
+	MinParams: 1,
+}
+
+
 var rebootApp = action.Action{
 	Name: "rebootapp",
 	Forward: func(ctx action.FWContext) (action.Result, error) {
-		var app global.DeepAssembly
+		var app handlers.AssemblyWithComponents
 		switch ctx.Params[0].(type) {
-		case global.DeepAssembly:
-			app = ctx.Params[0].(global.DeepAssembly)
-		case *handlers.DeepAssembly:
-			app = *ctx.Params[0].(*handlers.DeepAssembly)
+		case handlers.AssemblyWithComponents:
+			app = ctx.Params[0].(handlers.AssemblyWithComponents)
+		case *handlers.AssemblyWithComponents:
+			app = *ctx.Params[0].(*handlers.AssemblyWithComponents)
 		default:
-			return nil, errors.New("First parameter must be App or *handlers.DeepAssembly.")
+			return nil, errors.New("First parameter must be App or *handlers.AssemblyWithComponents.")
 		}
 		return TorpedoCommandExecutor("reboot", &app)
 	},
@@ -203,14 +332,14 @@ var rebootApp = action.Action{
 var restartApp = action.Action{
 	Name: "restartapp",
 	Forward: func(ctx action.FWContext) (action.Result, error) {
-		var app global.DeepAssembly
+		var app handlers.AssemblyWithComponents
 		switch ctx.Params[0].(type) {
-		case global.DeepAssembly:
-			app = ctx.Params[0].(global.DeepAssembly)
-		case *handlers.DeepAssembly:
-			app = *ctx.Params[0].(*handlers.DeepAssembly)
+		case handlers.AssemblyWithComponents:
+			app = ctx.Params[0].(handlers.AssemblyWithComponents)
+		case *handlers.AssemblyWithComponents:
+			app = *ctx.Params[0].(*handlers.AssemblyWithComponents)
 		default:
-			return nil, errors.New("First parameter must be App or *handlers.DeepAssembly.")
+			return nil, errors.New("First parameter must be App or *handlers.AssemblyWithComponents.")
 		}
 		return CommandExecutor("restart", &app)
 	},
@@ -223,14 +352,14 @@ var restartApp = action.Action{
 var stopApp = action.Action{
 	Name: "stopapp",
 	Forward: func(ctx action.FWContext) (action.Result, error) {
-		var app global.DeepAssembly
+		var app handlers.AssemblyWithComponents
 		switch ctx.Params[0].(type) {
-		case global.DeepAssembly:
-			app = ctx.Params[0].(global.DeepAssembly)
-		case *handlers.DeepAssembly:
-			app = *ctx.Params[0].(*handlers.DeepAssembly)
+		case handlers.AssemblyWithComponents:
+			app = ctx.Params[0].(handlers.AssemblyWithComponents)
+		case *handlers.AssemblyWithComponents:
+			app = *ctx.Params[0].(*handlers.AssemblyWithComponents)
 		default:
-			return nil, errors.New("First parameter must be App or *handlers.DeepAssembly.")
+			return nil, errors.New("First parameter must be App or *handlers.AssemblyWithComponents.")
 		}
 
 		return CommandExecutor("stop", &app)
@@ -244,18 +373,44 @@ var stopApp = action.Action{
 var restartComponent = action.Action{
 	Name: "restartcomponent",
 	Forward: func(ctx action.FWContext) (action.Result, error) {
-		var app global.Component
+		var app handlers.Component
 		switch ctx.Params[0].(type) {
-		case global.Component:
-			app = ctx.Params[0].(global.Component)
-		case *global.Component:
-			app = *ctx.Params[0].(*global.Component)
+		case handlers.Component:
+			app = ctx.Params[0].(handlers.Component)
+		case *handlers.Component:
+			app = *ctx.Params[0].(*handlers.Component)
 		default:
-			return nil, errors.New("First parameter must be App or *global.Component.")
+			return nil, errors.New("First parameter must be App or *handlers.Component.")
 		}
 		ctype := strings.Split(app.ToscaType, ".")
 		app.Command = "stop " + ctype[2] + "; " + "start " + ctype[2]
 		return ComponentCommandExecutor(&app)
+	},
+	Backward: func(ctx action.BWContext) {
+		log.Info("[%s] Nothing to recover")
+	},
+	MinParams: 1,
+}
+
+var buildApp = action.Action{
+	Name: "buildApp",
+	Forward: func(ctx action.FWContext) (action.Result, error) {
+		var app handlers.Component
+		switch ctx.Params[0].(type) {
+		case handlers.Component:
+			app = ctx.Params[0].(handlers.Component)
+		case *handlers.Component:
+			app = *ctx.Params[0].(*handlers.Component)
+		default:
+			return nil, errors.New("First parameter must be App or *handlers.Component.")
+		}
+		ctype := strings.Split(app.ToscaType, ".")
+		megam_home, perr := config.GetString("megam_home")
+		if perr != nil {
+			return nil, perr
+		}
+		app.Command = megam_home + "megam_" + ctype[2] + "_builder/build.sh"
+		return RedeployCommandExecutor(&app)
 	},
 	Backward: func(ctx action.BWContext) {
 		log.Info("[%s] Nothing to recover")
